+++ R/blue.r
@@ -2,9 +2,9 @@
 #'
 #' Default defined according to "ISO".
 #'
-#' @param std a character string "ISO", "Sellaro" (plant biology), or "RS"
-#'   (remote sensing), or Landsat imagers, "LandsatTM", "LandsatETM", or
-#'   "LandsatOLI".
+#' @param std a character string "ISO", "Sellaro" (plant biology), "Broad"
+#'   (plant biology) or "RS" (remote sensing), or Landsat imagers, "LandsatTM",
+#'   "LandsatETM", or "LandsatOLI".
 #'
 #' @return a waveband object wavelength defining a wavelength range.
 #'
@@ -29,6 +29,11 @@ Blue <- function(std = "ISO") {
   } else if (std == "Sellaro") {
     new_waveband(420,
                  490,
+                 wb.name = paste(label, std, sep = "."),
+                 wb.label = label)
+  } else if (std == "broad") {
+    new_waveband(400,
+                 500,
                  wb.name = paste(label, std, sep = "."),
                  wb.label = label)
   } else if (std == "LandsatOLI" || std == "RS") {